Guardando manejo excepciones capacitaciones y recuperar contraseña

+++ src/app/capacitaciones/ver-capacitaciones/ver-capacitaciones.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Course } from 'src/app/model/course';
-import { ActivatedRoute, Router } from '@angular/router';
 import { CapacitacionesService } from 'src/app/service/capacitaciones.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-ver-capacitaciones',
@@ -9,23 +9,35 @@ import { CapacitacionesService } from 'src/app/service/capacitaciones.service';
   styleUrls: ['./ver-capacitaciones.component.css']
 })
 export class VerCapacitacionesComponent {
-  courseList:Course [] = [];
-  constructor(
 
-    private route: ActivatedRoute,
-    private router: Router,
-    private courseService: CapacitacionesService,
-  ) { }
+  courseList: Course[] = [];
+
+  constructor(private courseService: CapacitacionesService) { }
 
   ngOnInit(): void {
-    this.courseService.findAll().subscribe(course => this.courseList = course)
-    console.log(this.courseList);
-  }
-  deleteCourse(id: number){
-    this.courseService.deleteCourse(id);
-    
-    const index = this.courseList.findIndex((plan) => plan.id === id);
-    this.courseList.splice(index, 1);
+    this.courseService.findAll().subscribe(course => {
+      this.courseList = course;
+    });
   }
 
+  deleteCourse(id: number) {
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "¡No podrás revertir esto!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminarlo'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.courseService.deleteCourse(id);
+        
+        const index = this.courseList.findIndex((course) => course.id === id);
+        this.courseList.splice(index, 1);
+
+        Swal.fire('Eliminado', 'El curso ha sido eliminado', 'success');
+      }
+    });
+  }
 }
